Create flags.lua, s2roomctx.lua and touchups.lua Move flag setting related methods into flags.lua Move generation methods into roomgen.lua Move onlevel methods into touchups.lua
